yyyy-MM-dd formatted date added

+++ src/main/java/Functions/CreateJsonFileWithMap.java
@@ -5,6 +5,8 @@ import com.google.gson.Gson;
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -23,7 +25,25 @@ public class CreateJsonFileWithMap {
             int size = Math.min(jsonValue.size(), headers.size());
             Map<String, Object> jsonObject = new HashMap<>();
             for (int i = 0; i < size; i++) {
-                jsonObject.put(headers.get(i), jsonValue.get(i));
+
+                HashMap<String, Object> isDate = GetValidDate.isDate(jsonValue.get(i));
+                boolean isValidDate = (boolean) isDate.get("isDate");
+
+                if (isValidDate) {
+
+                    /*Note
+                     * Dynamically Check this text is date or not
+                     * */
+                    LocalDate date = (LocalDate) isDate.get("date");
+                    DateTimeFormatter formatters = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+                    String text = date.format(formatters);
+                    LocalDate parsedDate = LocalDate.parse(text, formatters);
+                    jsonObject.put(headers.get(i), parsedDate.toString());
+
+                } else {
+                    jsonObject.put(headers.get(i), jsonValue.get(i));
+                }
+
             }
 
             jsonArray.add(jsonObject);
@@ -31,6 +51,7 @@ public class CreateJsonFileWithMap {
         });
         return jsonArray;
     }
+
 
     public static void createJson(List<String> headers, List<String> value, char delimiter) {
 
@@ -40,18 +61,20 @@ public class CreateJsonFileWithMap {
 
             File file = new File("result.jsonl");
             //ensure new result file
-            file.delete();
+            if (file.exists()) {
+                file.delete();
+            }
             file.createNewFile();
 
             jsonArray.forEach(o -> {
                 try {
-//                    JSONObject json = new JSONObject();
-//                    json.putAll(o);
-                    Gson gson = new Gson();
-                    String s = gson.toJson(o);
+
+                    /*append true means save with previous data*/
                     FileWriter fileW = new FileWriter(file.getAbsolutePath(), true);
-                    fileW.write(s);
+
+                    fileW.write(new Gson().toJson(o));
                     fileW.write(System.lineSeparator());
+
                     fileW.close();
 
                 } catch (IOException e) {
